Update Solar_Shield2.stl
made the sides a flat angle instead of sloped instead of rounded so that the printer will stop failing at an angle shifting layer. Also reduced the wall thickness to 2mm instead of 3 mm since it does not need to be that thick. Overall the print time was reduced by 43 min